Add report generator
This adds the report generator.  The report generator queries Github
for open pull requests on the designated repositories; it then
generates a textual report listing each open pull request, in order
of its age, as well as summary data over all repositories.  The
report generator function is usable on its own, as well as via the
CLI.  If used via the CLI, the report generator provides selectable
verbosity, and is capable of prompting for Github API token or
password.