Emit more debug messages
When generating the report to standard output, we don't need any
special status messages, but when generating the report to a file,
it would be handy to know what tugboat is doing.  This adds some
extra status messages in the case of normal or verbose output,
when the output is directed to a file.